reffactoring and extend library to use different providers
