+++ src/providers/fastnear/client.rs
@@ -1,0 +1,1 @@
+
