+++ src/providers/fastnear/mod.rs
@@ -1,0 +1,3 @@
+pub mod client;
+pub mod fetchers;
+pub mod types;
